Datetime function is moved to commons
